Additionally, a one-time service component has been added, the SubscriptionServiceModal has been renamed to OneTimeServiceModal and components to handle one-time and subscription services were extracted out from the old SubscriptionServiceModal into independent forms. This was done to offer better structured, more reusable and more testable components. Also, FormInput component was modified to allow for custom FormLabelProps, providing more flexibility in label customization.
SubscriptionServiceModal was earlier tightly coupled, and handled both one-time and subscription services. By having separate forms, it provides a clear distinction between the two types of service and allows more flexibility and control. Also, renaming SubscriptionServiceModal to ServiceModal makes it clear that it can encompass any kind of service, and not just subscription ones.
