+++ packages/apps/spaces/ui/form/Input/FormInput.tsx
@@ -3,7 +3,12 @@
 import { forwardRef } from 'react';
 import { useField } from 'react-inverted-form';
 
-import { FormLabel, FormControl, VisuallyHidden } from '@chakra-ui/react';
+import {
+  FormLabel,
+  FormControl,
+  VisuallyHidden,
+  FormLabelProps,
+} from '@chakra-ui/react';
 
 import { Input, InputProps } from './Input';
 
@@ -12,18 +17,29 @@ interface FormInputProps extends InputProps {
   formId: string;
   label?: string;
   isLabelVisible?: boolean;
+  labelProps?: FormLabelProps;
 }
 
 //todo add visually hidden label - accessibility
 
 export const FormInput = forwardRef(
-  ({ name, formId, label, isLabelVisible, ...props }: FormInputProps, ref) => {
+  (
+    {
+      name,
+      formId,
+      label,
+      isLabelVisible,
+      labelProps,
+      ...props
+    }: FormInputProps,
+    ref,
+  ) => {
     const { getInputProps } = useField(name, formId);
 
     return (
       <FormControl>
         {isLabelVisible ? (
-          <FormLabel>{label}</FormLabel>
+          <FormLabel {...labelProps}>{label}</FormLabel>
         ) : (
           <VisuallyHidden>
             <FormLabel>{label}</FormLabel>
